backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
PyMuPDF
streamlit
